+++ CHANGES.rst
@@ -1,9 +1,9 @@
 Version 0.11.0   NOT RELEASED
 ----------------------------
 
-* FEATURE: Added default filter method lookup for MethodFilter
+* FEATURE: Added default filter method lookup for MethodFilter #222
 
-
+* FEATURE: Added support for yesterday in daterangefilter #234
 
 
 Version 0.10.0 (2015-05-13)