+++ CHANGES.rst
@@ -7,6 +7,8 @@ Version 0.11.0   NOT RELEASED
 
 * FEATURE: Created Filter for NumericRange. #236
 
+* FEATURE: Added Date/time range filters #215
+
 
 Version 0.10.0 (2015-05-13)
 ---------------------